Created reservation edit page (not functional yet)
