+++ app/src/main/java/com/example/jdm_app/domain/RentCondition.kt
@@ -2,6 +2,7 @@ package com.example.jdm_app.domain
 
 import com.example.jdm_app.adapter.DateJsonAdapter
 import com.squareup.moshi.Json
+import com.squareup.moshi.JsonAdapter
 import java.io.Serializable
 import java.util.*
 
@@ -9,7 +10,7 @@ data class RentCondition(
     @Json(name = "id")
     val id: Int? = null,
 
-    @JsonAdapter(DateJsonAdapter::class)
+    @JsonAdapter(DateJsonAdapter::toJson)
     var rentDate: Date? = null,
 
     @Json(name = "postalCode")
