+++ app/src/main/res/layout/reservation_edit.xml
@@ -64,6 +64,7 @@
         android:id="@+id/edit_text_house_number"
         android:hint="@string/house_number"
         android:inputType="text"
+        android:minHeight="48dp"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         tools:ignore="TextFields" />
